legg til tester og fiks bug i beregnOrganisasjoner
bugen gjorde at brukere som hadde tilgang til altinn i organisasjonen høyere opp i treet ikke fikk
tilgang til virksomhet basert på digisyfo lenger ned i treet.
