+++ package.json
@@ -29,8 +29,8 @@
     "@graphql-codegen/typescript": "^4.1.3",
     "@graphql-codegen/typescript-operations": "^4.4.1",
     "@graphql-codegen/typescript-react-apollo": "^4.3.2",
-    "@testing-library/jest-dom": "^6.4.6",
-    "@testing-library/react": "^16.0.0",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.2.0",
     "@types/jest-axe": "^3.5.9",
     "@types/node": "^22.13.1",
     "@types/react": "18.3.3",
