+++ vite-env.d.ts
@@ -1,5 +1,6 @@
 /// <reference types="vite/client" />
 
 declare const __BASE_PATH__: string;
+declare const __BUILD_TIMESTAMP__: string;
 
 declare module '*.graphql';
